refactor(extension): 迁 runDateRangeDriverCDP 到 cdp-drivers/daterange #0000
- dom.ts L1414-1794（parseYMDLocal + runDateRangeDriverCDP 共 381 行） → adapter/cdp-drivers/daterange.ts
- dom.ts 2127 → 1746（净减 380 行，新增 1 行 import）
- cdp-drivers/index.ts re-export runDateRangeDriverCDP
- 函数体逐行原样迁移，调用点 dom.ts L913 不变（debuggerMgr 已通过 opts 注入）
- pageQuery / clickBBox alias 一并迁入 driver 内部
- build 通过，test 基线 6 failed / 150 passed 0 回归，I1 depcruise invariant 绿
