+++ packages/extension/src/adapter/cdp-drivers/index.ts
@@ -1,4 +1,5 @@
 // COMMIT handler 路由的 CDP driver 集合。
 // T1.7a / T1.7b / T1.7c 逐步迁入。
 
-export {}; // T1.7a 起填充
+export { runDateRangeDriverCDP } from "./daterange.js";
+// runCascaderDriverCDP / runTimePickerDriverCDP 在 T1.7b / T1.7c 加入
